Working application code!!!!!!!!
- I'm still trying to figure out how the benefits of "prepared" queries shake
out — it looks like doing a `.bind()` mutates the query itself in a
non-reversible way, so you can't hold onto the query struct after using it once.
So, is this all happening in the statement cache of the database itself???
Unclear! Or possibly, known by all but me such that it needs no saying. 🪼

- `#[derive(FromRow)]` makes this SOOOOOOOOO much less verbose than rusqlite.
Even without doing compile-time query checking thru the macros, I think that's
the killer feature.